refactor(rolldown_plugin_web_worker_post): tweak code organization
